[KOGITO-8198] Fixing problem in avro schema for ints (#2587)
* [KOGITO-8198] Fixing problem in avro schema for ints

* [KOGITO-8198] Sonar issues
